mb/google/calypso: Enable boot media detection for Calypso
Add a Calypso-specific devicetree under DEVICETREE with FW_CONFIG
fields that differ from the default devicetree for boot media
detection.

Move fw_config_override() from mainboard.c to fw_config_calypso.c,
which is compiled only for Calypso.

Populate FW_CONFIG with the detected storage media so the correct
boot storage interface is enumerated during boot.

TEST=Build and boot google/calypso

Reviewed-on: https://review.coreboot.org/c/coreboot/+/95042
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
