+++ src/mainboard/google/calypso/Makefile.mk
@@ -11,8 +11,10 @@ bootblock-y += bootblock.c
 romstage-y += romstage.c
 
 romstage-y += charging.c
+romstage-$(CONFIG_SOC_QUALCOMM_CDT) += fw_config_calypso.c
 
 ramstage-y += charging.c
+ramstage-$(CONFIG_SOC_QUALCOMM_CDT) += fw_config_calypso.c
 
 ramstage-y += mainboard.c
